Contractor successfully fulfills a wish:
I can't say it fulfills a contract yet because one has not been
provided. My development strategy is to test drive the Contractor class
via the examples, and the examples will be developed with TDD. So there
my not be many tests directly testing the core classes, but they will be
tested indirectly through the examples that use them. In this case, I
started on a DCI library for PHP as an example of something that uses
the core classes. With each commit, the example I am working on may or
may not have achieved the next level of usefulness, but they are heading
toward something more useful, and their tests will pass, thereby
indirectly driving the development and testing of the core classes they
are an example of using.

This is a way of applying the transformation priority premise to
developing the core classes. Theoretically, as long as I do TDD,
constantly refactor to have clean code, and follow the transformation
priorities, I should end up with good algorithms that work, are
efficient, and have clean code. Of course that is a very idealistic view
of things and this is really just an experiment to test out that
hypothesis and see where it takes us.