RolePlayer throws a nice exception if the injected method is missing:
And yes this is happening via the Contractor. But at this point it is
just a cheap magic trick. That is ok though, thus is the nature of TDD
with the Bob's three laws and the transformation priorities. It will
only ever be as intelligent as it needs to be to satisfy the cases that
have been tested. Hopefully amazing algorithms will magically appear
without trying like what happens when qicksort appears in a sort method.