Work in progress commit:
There was duplication in the contracts between inject() and call().
I defined a "relationship" that is actually just a condition.
Conceptually this is still a contract, but it defies the convention of
contracts being preconditions, postconditions, and invariants. This is a
more complex contract that expresses a relationship between 2 methods.
Since it is returned by a method, both methods that share this
relationship call it, thus avoiding the duplication.